src/backend/requirements.txt
astroid==1.6.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==0.12.2
gunicorn==19.9.0
grequests==0.3.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.8.2
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
